#7058 Remove unneeded changes
