+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/picking/service/ForwardingPackingItem.java
@@ -1,16 +1,17 @@
 package de.metas.picking.service;
 
+import java.util.Set;
+import java.util.function.Predicate;
+
+import org.adempiere.warehouse.WarehouseId;
+import org.compiere.model.I_C_UOM;
+
 import de.metas.bpartner.BPartnerId;
 import de.metas.bpartner.BPartnerLocationId;
 import de.metas.handlingunits.HUPIItemProductId;
 import de.metas.inoutcandidate.ShipmentScheduleId;
 import de.metas.product.ProductId;
 import de.metas.quantity.Quantity;
-import org.adempiere.warehouse.WarehouseId;
-import org.compiere.model.I_C_UOM;
-
-import java.util.Set;
-import java.util.function.Predicate;
 
 /*
  * #%L
@@ -92,11 +93,6 @@ abstract class ForwardingPackingItem implements IPackingItem
 	public void addParts(final IPackingItem packingItem)
 	{
 		getDelegate().addParts(packingItem);
-	}
-
-	public final IPackingItem addPartsAndReturn(final IPackingItem packingItem)
-	{
-		return getDelegate().addPartsAndReturn(packingItem);
 	}
 
 	@Override
