+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/picking/service/PackingItem.java
@@ -3,7 +3,19 @@
  */
 package de.metas.picking.service;
 
+import java.util.Set;
+import java.util.function.Predicate;
+
+import javax.annotation.Nullable;
+
+import org.adempiere.exceptions.AdempiereException;
+import org.adempiere.util.lang.impl.TableRecordReference;
+import org.adempiere.warehouse.WarehouseId;
+import org.compiere.model.I_C_UOM;
+import org.compiere.util.Util;
+
 import com.google.common.collect.ImmutableSet;
+
 import de.metas.bpartner.BPartnerId;
 import de.metas.bpartner.BPartnerLocationId;
 import de.metas.handlingunits.HUPIItemProductId;
@@ -14,15 +26,6 @@ import de.metas.quantity.Quantity;
 import de.metas.util.Check;
 import de.metas.util.Services;
 import lombok.NonNull;
-import org.adempiere.exceptions.AdempiereException;
-import org.adempiere.util.lang.impl.TableRecordReference;
-import org.adempiere.warehouse.WarehouseId;
-import org.compiere.model.I_C_UOM;
-import org.compiere.util.Util;
-
-import javax.annotation.Nullable;
-import java.util.Set;
-import java.util.function.Predicate;
 
 /**
  * Item to be packed.
@@ -42,7 +45,7 @@ final class PackingItem implements IPackingItem
 
 		this.parts = parts.copy();
 		this.groupingKey = parts.mapReduce(PackingItem::computeGroupingKey).get();
-		uom = parts.getCommonUOM();
+		uom = parts.mapReduce(part -> part.getQty().getUOM()).get();
 	}
 
 	PackingItem(@NonNull final PackingItem copyFromItem)
@@ -228,13 +231,6 @@ final class PackingItem implements IPackingItem
 	{
 		addParts(packingItem.getParts());
 	}
-	@Override
-	public final IPackingItem addPartsAndReturn(final IPackingItem packingItem)
-	{
-		addParts(packingItem.getParts());
-		return this;
-	}
-
 
 	@Override
 	public final void setPartsFrom(final IPackingItem packingItem)
