+++ backend/metasfresh-webui-api/src/main/java/de/metas/ui/web/picking/pickingslot/process/WEBUI_Picking_PickQtyToNewHU.java
@@ -1,6 +1,22 @@
 package de.metas.ui.web.picking.pickingslot.process;
 
+import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_MISSING_SOURCE_HU;
+import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_SELECT_PICKING_SLOT;
+
+import java.math.BigDecimal;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.Properties;
+
+import org.adempiere.exceptions.AdempiereException;
+import org.adempiere.warehouse.LocatorId;
+import org.adempiere.warehouse.WarehouseId;
+import org.adempiere.warehouse.api.IWarehouseBL;
+import org.compiere.model.I_C_UOM;
+import org.springframework.beans.factory.annotation.Autowired;
+
 import com.google.common.collect.ImmutableList;
+
 import de.metas.handlingunits.HUPIItemProductId;
 import de.metas.handlingunits.HuId;
 import de.metas.handlingunits.IHUPIItemProductBL;
@@ -11,10 +27,15 @@ import de.metas.handlingunits.model.I_M_HU_PI;
 import de.metas.handlingunits.model.I_M_HU_PI_Item_Product;
 import de.metas.handlingunits.model.I_M_ShipmentSchedule;
 import de.metas.handlingunits.model.X_M_HU;
+import de.metas.handlingunits.picking.PickingCandidateService;
+import de.metas.handlingunits.picking.requests.AddQtyToHURequest;
 import de.metas.handlingunits.report.HUReportService;
 import de.metas.handlingunits.report.HUToReportWrapper;
 import de.metas.i18n.ITranslatableString;
 import de.metas.inoutcandidate.api.IShipmentScheduleEffectiveBL;
+import de.metas.order.DeliveryRule;
+import de.metas.picking.api.PickingConfigRepository;
+import de.metas.picking.api.PickingSlotId;
 import de.metas.process.IProcessDefaultParameter;
 import de.metas.process.IProcessDefaultParametersProvider;
 import de.metas.process.IProcessPrecondition;
@@ -30,19 +51,6 @@ import de.metas.ui.web.window.datatypes.LookupValue.IntegerLookupValue;
 import de.metas.ui.web.window.datatypes.LookupValuesList;
 import de.metas.util.Services;
 import lombok.NonNull;
-import org.adempiere.exceptions.AdempiereException;
-import org.adempiere.warehouse.LocatorId;
-import org.adempiere.warehouse.WarehouseId;
-import org.adempiere.warehouse.api.IWarehouseBL;
-import org.compiere.model.I_C_UOM;
-
-import java.math.BigDecimal;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.Properties;
-
-import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_MISSING_SOURCE_HU;
-import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_SELECT_PICKING_SLOT;
 
 /*
  * #%L
@@ -78,6 +86,12 @@ public class WEBUI_Picking_PickQtyToNewHU
 		extends WEBUI_Picking_With_M_Source_HU_Base
 		implements IProcessPrecondition, IProcessDefaultParametersProvider
 {
+	@Autowired	
+	private PickingCandidateService pickingCandidateService;	
+
+	@Autowired	
+	private PickingConfigRepository pickingConfigRepo;
+	
 	private final IWarehouseBL warehouseBL = Services.get(IWarehouseBL.class);
 
 	private static final String PARAM_M_HU_PI_Item_Product_ID = I_M_HU_PI_Item_Product.COLUMNNAME_M_HU_PI_Item_Product_ID;
@@ -113,13 +127,9 @@ public class WEBUI_Picking_PickQtyToNewHU
 
 	protected Optional<ProcessPreconditionsResolution> checkValidSelection()
 	{
-		if (getParentViewRowIdsSelection() == null)
-		{
-			return Optional.of(ProcessPreconditionsResolution.rejectBecauseNoSelection());
-		}
-
-		if (getSelectedRowIds().isMoreThanOneDocumentId()
-	        ||  getParentViewRowIdsSelection().getRowIds().isMoreThanOneDocumentId() )
+		if (!getSelectedRowIds().isSingleDocumentId()
+				|| (getParentViewRowIdsSelection() != null
+						&& getParentViewRowIdsSelection().getRowIds().isMoreThanOneDocumentId()))
 		{
 			return Optional.of(ProcessPreconditionsResolution.rejectBecauseNotSingleSelection());
 		}
@@ -133,6 +143,11 @@ public class WEBUI_Picking_PickQtyToNewHU
 		return Optional.empty();
 	}
 
+	protected boolean isForceDelivery()	
+	{	
+		return DeliveryRule.ofCode(getCurrentShipmentSchedule().getDeliveryRule()).isForce();	
+	}
+	
 	protected Quantity getQtyToPack()
 	{
 		final I_C_UOM uom = getCurrentShipmentScheuduleUOM();
@@ -170,6 +185,23 @@ public class WEBUI_Picking_PickQtyToNewHU
 		final HUToReportWrapper huToReport = HUToReportWrapper.of(hu);
 		final boolean onlyIfAutoPrintIsEnabled = true;
 		huReportService.printPickingLabel(huToReport, onlyIfAutoPrintIsEnabled);
+	}
+	
+	protected Quantity pickHUsAndPackTo(@NonNull final ImmutableList<HuId> huIdsToPick, @NonNull final Quantity qtyToPack, @NonNull final HuId packToHuId)	
+	{	
+		final boolean allowOverDelivery = pickingConfigRepo.getPickingConfig().isAllowOverDelivery();	
+
+		final PickingSlotRow pickingSlotRow = getSingleSelectedRow();	
+		final PickingSlotId pickingSlotId = pickingSlotRow.getPickingSlotId();	
+
+		return pickingCandidateService.addQtyToHU(AddQtyToHURequest.builder()	
+				.qtyToPack(qtyToPack)	
+				.packToHuId(packToHuId)	
+				.sourceHUIds(huIdsToPick)	
+				.pickingSlotId(pickingSlotId)	
+				.shipmentScheduleId(getCurrentShipmentScheduleId())	
+				.allowOverDelivery(allowOverDelivery)	
+				.build());	
 	}
 
 	@NonNull
